chore(release): bump version to 0.1.2-rc.2 and modularize provider constants
