+++ apps/web/package.json
@@ -1,6 +1,6 @@
 {
     "name": "web",
-    "version": "0.1.2-rc.1",
+    "version": "0.1.2-rc.2",
     "type": "module",
     "private": true,
     "scripts": {
