+++ packages/constants/src/index.ts
@@ -1,4 +1,4 @@
-export * from "./providers.js";
+export * from "./providers/index.js";
 export * from "./oauth.js";
 export * from "./seed.js";
 export * from "./version.js";
