+++ packages/constants/src/providers/anthropic.ts
@@ -1,0 +1,1 @@
+export const ANTHROPIC_BASE_URL = "https://api.anthropic.com/v1";
