+++ packages/constants/src/providers/openai.ts
@@ -1,0 +1,3 @@
+export const OPENAI_BASE_URL = "https://api.openai.com/v1";
+export const CODEX_BASE_URL = "https://chatgpt.com/backend-api/codex/responses";
+export const CODEX_MODELS_URL = "https://chatgpt.com/backend-api/codex/models";
