+++ packages/constants/src/version.ts
@@ -1,4 +1,4 @@
-export const GLOBAL_VERSION = "0.1.2-rc.1";
+export const GLOBAL_VERSION = "0.1.2-rc.2";
 export const APP_VERSION = GLOBAL_VERSION;
 export const API_VERSION = GLOBAL_VERSION;
 export const CLI_VERSION = GLOBAL_VERSION;
